backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-CORS>=3.0
python-dotenv>=0.21
requests>=2.28